Working implementation of Pd externals for O2. Note that in this implementation, o2receive objects should be initialized AFTER o2ensemble is initialized. You can connect loadbang to o2receive, which will try to create a service and message handler if they are not already created due to an earlier failure. The bang is ignored if o2receive is receiving already. Test on MacOS, which also passes all regression tests (test/rt.py). There were previous errors discovered around the MTQQ service connection being confused with TCP connections to other O2 processes.

+++ src/o2lite.c
@@ -799,8 +799,8 @@ void read_from_tcp()
                 while (tcp_msg_got < tcp_in_msg->length) {
                     int togo = tcp_in_msg->length - tcp_msg_got;
                     if (togo > capacity) togo = capacity;
-                    n = recvfrom(tcp_sock, PTR(&tcp_in_msg->misc), togo, 
-                                 0, NULL, NULL);
+                    n = (long) recvfrom(tcp_sock, PTR(&tcp_in_msg->misc), togo,
+                                        0, NULL, NULL);
                     if (n <= 0) {
                         goto error_exit;
                     }
@@ -911,7 +911,7 @@ void o2l_send()
     if (parse_error || tcp_sock == INVALID_SOCKET) {
         return;
     }
-    out_msg->length = o2lswap32(out_msg_cnt - sizeof out_msg->length);
+    out_msg->length = (int32_t) o2lswap32(out_msg_cnt - sizeof out_msg->length);
     if (out_msg->misc & o2lswap32(O2_TCP_FLAG)) {
         send(tcp_sock, outbuf, out_msg_cnt, 0);
     } else {
